+++ appObjects/FlatCAMDocument.py
@@ -164,14 +164,14 @@ class DocumentObject(FlatCAMObj):
 
         # try to not add too many times a tab that it is already installed
         for idx in range(self.app.ui.plot_tab_area.count()):
-            if self.app.ui.plot_tab_area.widget(idx).objectName() == self.options['name']:
+            if self.app.ui.plot_tab_area.widget(idx).objectName() == self.options['name'] + "_editor_tab":
                 tab_here = True
                 break
 
         # add the tab if it is not already added
         if tab_here is False:
             self.app.ui.plot_tab_area.addTab(self.document_editor_tab, '%s' % _("Document Editor"))
-            self.document_editor_tab.setObjectName(self.options['name'])
+            self.document_editor_tab.setObjectName(self.options['name'] + "_editor_tab")
 
         # Switch plot_area to CNCJob tab
         self.app.ui.plot_tab_area.setCurrentWidget(self.document_editor_tab)